Stuff you can mine
This commit contains all of the .ENTITY.MBIN files for things that you
can mine (for the most part).  By editing these files I should be able
to specifically reduce the occurrence of certain resources.
- Can change RARITY of an entity
- Can change the amount of Substance that is given by an entity
- Can Change the Loot amount given by an entity.
- Can change the health of entities.
- Can change if mining an entity Increases the Wanted level!!!
- Can change how much an entity increases your wanted level by if you're
currently wanted!!!
- Can change ScannableComponentData
- Pretty sure this is how close and for how long you have to scan the
item to log it as a discovery, **not** whether or not it shows up when
you do a scan for stuff.  Still need to test.
- Interestingly ResourceRock.Entity gives 10 Fuel while
MediumRockScan.Entity gives 50 Commodity.
- LargeRockEntity gives 50 Commodity, ResourceRockLarge gives 20 Fuel
and ResourceRockShard gives 40 fuel.
- Will need to further explore to figure out what is what.

##Lots of info in here and it's gonna take awhile to sort through it
all.